tlscfg: drop setting CurvePreferences by default
A few people have had problems with this, and it's complicated to
determine why (past my usual recommendation of trying to drop this
field).

NSA's P curves are only ever suspected of being backdoored, but
apparently nobody _believes_ it. Part of the problem with P256 is that
it's complicated to implement, but Go's security is pretty reliable.

We'll opt into the default P curves in addition to x25519.